issue #729 consolidate commons
