+++ commons/src/main/scala/za/co/absa/commons/CollectionImplicits.scala
