+++ commons/src/main/scala/za/co/absa/commons/EnumUtils.scala
