+++ commons/src/main/scala/za/co/absa/commons/ExtractorImplicits.scala
