+++ commons/src/main/scala/za/co/absa/commons/NamedEntity.scala
