+++ commons/src/main/scala/za/co/absa/commons/SplineTraversableExtension.scala
