+++ commons/src/main/scala/za/co/absa/commons/ThrowableImplicits.scala
