+++ commons/src/test/scala/za/co/absa/commons/CollectionImplicitsSpec.scala
