+++ commons/src/test/scala/za/co/absa/commons/EnumUtilsSpec.scala
