+++ core/src/main/scala/za/co/absa/spline/HierarchicalObjectFactory.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 ABSA Group Limited
+ * Copyright 2023 ABSA Group Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package za.co.absa.commons
+package za.co.absa.spline
 
 import org.apache.commons.configuration.Configuration
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SparkSession
-import za.co.absa.commons.HierarchicalObjectFactory.ClassName
 import za.co.absa.commons.config.ConfigurationImplicits.ConfigurationRequiredWrapper
+import za.co.absa.spline.HierarchicalObjectFactory.ClassName
 
 import java.lang.reflect.InvocationTargetException
 import scala.reflect.ClassTag
