+++ core/src/main/scala/za/co/absa/spline/agent/AgentBOM.scala
@@ -18,7 +18,7 @@ package za.co.absa.spline.agent
 
 import org.apache.commons.configuration.{CompositeConfiguration, Configuration}
 import org.apache.spark.sql.SparkSession
-import za.co.absa.commons.HierarchicalObjectFactory
+import za.co.absa.spline.HierarchicalObjectFactory
 import za.co.absa.spline.agent.AgentConfig.ConfProperty
 import za.co.absa.spline.harvester.IdGenerator.UUIDVersion
 import za.co.absa.spline.harvester.conf.{SQLFailureCaptureMode, SplineMode}
