+++ core/src/main/scala/za/co/absa/spline/harvester/conf/Spline05ConfigurationAdapter.scala
@@ -17,7 +17,7 @@
 package za.co.absa.spline.harvester.conf
 
 import org.apache.commons.configuration.Configuration
-import za.co.absa.commons.HierarchicalObjectFactory.ClassName
+import za.co.absa.spline.HierarchicalObjectFactory.ClassName
 import za.co.absa.spline.agent.AgentConfig.ConfProperty._
 import za.co.absa.spline.harvester.conf.Spline05ConfigurationAdapter._
 import za.co.absa.spline.harvester.dispatcher.httpdispatcher.HttpLineageDispatcherConfig._
