+++ core/src/test/scala/za/co/absa/spline/agent/AgentBOMSpec.scala
@@ -21,7 +21,8 @@ import org.apache.spark.sql.SparkSession
 import org.scalatest.OptionValues._
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import za.co.absa.commons.{HierarchicalObjectFactory, NamedEntity}
+import za.co.absa.commons.NamedEntity
+import za.co.absa.spline.HierarchicalObjectFactory
 import za.co.absa.spline.agent.AgentBOMSpec.{MockLineageDispatcher, MockPostProcessingFilter, SimpleConfig, createBOM}
 import za.co.absa.spline.agent.AgentConfig.ConfProperty
 import za.co.absa.spline.harvester.conf.{SQLFailureCaptureMode, SplineMode}
